Status page: Display total number of accounts
As described in #59, this publicized the current number of active accounts/pantries.
I don't know how big Pantry is or how much traffic you get; it might make sense to cache the total number of accounts, although I'm not familiar enough with Redis to know how to mark that cache as stale when Redis marks a key as expired (though it's probably not a huge deal if old accounts are included in the tally for a bit?).

Anyway, this seems like an intriguing project and I look forward to helping out more!