Allow Pantry to Run in Development Mode
We have to somehow create a "development mode" for the app. The problem
developers are currently facing is that they have to remove lines of code to
bypass the recaptcha because they do not have a valid secret key.

It would be useful if we somehow added a development mode to bypass the
recaptcha, & other prod-specific service functions.

* Adds an environment service to check if Pantry is running in development mode
* Adds a new env var in sample.env called NODE_ENV. If set to 'development' then
* Pantry will run in development mode.
* Preventing Pantry from publishing Slack messages, posting to the CRM and
  disabling the recaptcha when running in development mode

Resolves #90