Allow Users to Edit Pantry Contents via Dashboard
The dashboard as it stands now is very much a read-only application.
We should allow users who do not want to use the API to be able to manipulate
their pantry contents (baskets) via the dashboard. Users should be able to add,
remove and edit existing baskets via some form of JSON editor.

* Replace json-pretty with json-editor
* Abstracted backet logic to a new vue component
* Add ability to delete baskets
* Add ability to create new baskets
* Changed copy to reflect new features

Resolves #112