Remove Backend Assets from Dist Folder
We initally added the dist folder to the repo to speed up our deploys due to the
front end assets taking too long to compile. In doing so, we also included the
backend assets. These items can be discarded as they do not slow down deploys.

This change addresses the need by:
* Deleting backend folders
* Isolating the frontend assets through gitignore