Change Basket Modify Request Method to POST
Users should be able to update the contents of the basket entirely via
the dashboard. Currently, the dashboard uses a put to add/edit the
contents of the basket. However, users would probably like to have full
control and overwrite the entirety of a basket's content if they wish.

This change addresses the need by:
* Updating the request method from PUT to POST

Resolves #150