Improve Performance and Reduce Complexity
A bit of background info. Pantry is currently hosted on a AWS Lightsail
instance and has been seeing a large increase in traffic. Due to this,
performance issues have come to light and Pantry has suffered many
outages in the last few weeks. The memory usage of the application would
spike, causing the instance to encroach into the burstable zone. Once
there, Lightsail would restart.

The root issue was Logflare, which changed its pricing and blocked our
API requests, slowing down the thread for each request.

In order to speed up performance of the application we need to remove
Logflare, stop emailing users of each error, prevent accounts from
storing errors, and include a dist folder which speeds up deployment.